+++ SimpleChatMonitor/CHANGELOG.md
@@ -4,6 +4,7 @@
 - Implemented a max loss factor for the gambler
 - Implemented a voting system
 - Separated twitch bot functionality in to its own module
+- Added arkas sellout stream calculator
 
 ## [2.2.0] - 20-06-2022